refactor: 함수 네이밍 변경 (#11)
